service-workers/service-worker/navigation-preload/request-headers.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Navigation Preload request headers</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.sub.js"></script>
<script>
promise_test(async t => {
  const SCRIPT = 'resources/request-headers-worker.js';
  const SCOPE = 'resources/request-headers-scope.py';
  const registration =
      await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
  const worker = registration.installing;
  await wait_for_state(t, worker, 'activated');
  const frame = await with_iframe(SCOPE);

  add_completion_callback(async() => {
    if (registration)
      await registration.unregister();
    if (frame)
      frame.remove();
  })

  const headers = JSON.parse(frame.contentDocument.body.textContent);
  assert_true('SERVICE-WORKER-NAVIGATION-PRELOAD' in headers,
              'The Navigation Preload request must specify a ' +
              '"Service-Worker-Navigation-Preload" header.');
  assert_array_equals(headers['SERVICE-WORKER-NAVIGATION-PRELOAD'], ['hello'],
                      'The Navigation Preload request must specify the ' +
                      'correct value for the ' +
                      '"Service-Worker-Navigation-Preload" header.');
  assert_true('UPGRADE-INSECURE-REQUESTS' in headers,
              'The Navigation Preload request must specify an ' +
              '"Upgrade-Insecure-Requests" header.');
  assert_array_equals(headers['UPGRADE-INSECURE-REQUESTS'], ['1'],
                      'The Navigation Preload request must specify the ' +
                      'correct value for the ' +
                      '"Upgrade-Insecure-Requests" header.');
}, 'Navigation Preload request headers.');

</script>
